rearrange and make nested tables private functions
