+++ plugins/source/aws/resources/services/s3/bucket_grants.go
@@ -1,13 +1,17 @@
 package s3
 
 import (
+	"context"
+
+	"github.com/aws/aws-sdk-go-v2/service/s3"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/cloudquery/cloudquery/plugins/source/aws/client"
+	"github.com/cloudquery/cloudquery/plugins/source/aws/resources/services/s3/models"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
 )
 
-func BucketGrants() *schema.Table {
+func bucketGrants() *schema.Table {
 	return &schema.Table{
 		Name:        "aws_s3_bucket_grants",
 		Description: `https://docs.aws.amazon.com/AmazonS3/latest/API/API_Grant.html`,
@@ -41,4 +45,23 @@ func BucketGrants() *schema.Table {
 			},
 		},
 	}
+}
+func fetchS3BucketGrants(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	r := parent.Item.(*models.WrappedBucket)
+	svc := meta.(*client.Client).Services().S3
+	region := parent.Get("region").(*schema.Text)
+	if region == nil {
+		return nil
+	}
+	aclOutput, err := svc.GetBucketAcl(ctx, &s3.GetBucketAclInput{Bucket: r.Name}, func(options *s3.Options) {
+		options.Region = region.Str
+	})
+	if err != nil {
+		if client.IsAWSError(err, "NoSuchBucket") {
+			return nil
+		}
+		return err
+	}
+	res <- aclOutput.Grants
+	return nil
 }
